Create new lazy message types that wrap a *Frame
The wrapped classes are the same size as the pointer, and allow us to
add messageType-specific methods on top of a frame rather than adding
them to the frame directly.

In future, we may want to cache items in the wrapped class.
